v2.3.0 - Fix x40 volume stepping, 0.5 dB precision, upgrade libraries
- Fix x40 volume stepping: use native VUP/VDN (0.5 dB) instead of
  PVUP/PVDN which caused non-linear 2-3 dB jumps at different levels
- Add optimistic state updates for volume_up_step/volume_down_step
  so the UI responds instantly without waiting for TCP round-trip
- Widen parser to capture fractional dB values (e.g. -30.5)
- Change volume_db type from int to float throughout the data path
- Format float values with :g in _get_zone_command to preserve
  retry-registry key matching (Z1VOL-30 matches for both int and float)
- Fix Remote entity to use ucapi 0.7.0 constructor kwargs
  (simple_commands/ui_pages) instead of manual self.options override
- Upgrade ucapi-framework to 1.9.5, ucapi to 0.7.0

+++ uc_intg_anthemav/models.py
@@ -14,7 +14,7 @@ class ZoneState:
     """Represents the state of a single zone."""
 
     power: Optional[bool] = None
-    volume_db: Optional[int] = None
+    volume_db: Optional[float] = None
     volume_pct: Optional[int] = None
     muted: Optional[bool] = None
     input_number: Optional[int] = None
@@ -77,7 +77,7 @@ class ZonePower(ZoneMessage):
 class ZoneVolume(ZoneMessage):
     """Zone volume in dB (ZxVOL)."""
 
-    volume_db: int
+    volume_db: float
 
 
 @dataclass
